+++ packages/web-components/src/components/cbp-button/cbp-button.tsx
@@ -1,10 +1,10 @@
 import { Component, Prop, Element, Event, EventEmitter, Host, h } from '@stencil/core';
-import { setCSSProps, getElementAttrs } from '../../utils/utils';
+import { setCSSProps, getElementAttrs, createNamespaceKey } from '../../utils/utils';
 //import state from './store';
 
 /**
  * @slot - The button's label, which may contain markup such as icons, is placed in the default slot.
- * @slot - cbp-button-custom - Custom buttons/anchors may be slotted via named-slot, which prevents the web component from rendering its own tag. In such a case, component properties that render attributes directly onto the rendered button will have no effect, as they are expected to be supplied directly on the slotted custom element.
+ * @slot - cbp-button-custom - Custom buttons/anchors may be slotted via named-slot, which prevents the web component from rendering its own tag. In such a case, component properties that render attributes directly onto the rendered button will have no effect, as they are expected to be supplied directly on the slotted element.
  */
 @Component({
   tag: 'cbp-button',
@@ -16,7 +16,7 @@ export class CbpButton {
 
   private persistedAttrs: any;
 
-  @Element() host: HTMLElement;
+  @Element() host: HTMLCbpButtonElement;
 
   /** Specifies whether the button is a true button element or "link button." */
   @Prop() tag: 'button' | 'a' = 'button';
@@ -29,6 +29,8 @@ export class CbpButton {
   /** Specifies a variant of the buttons, such as square for buttons with only an icon and call-to-action button. */
   @Prop({ reflect: true }) variant: 'square' | 'cta';
 
+  /** Optionally specify the ID of the control here, which is used to generate related pattern node IDs and associate everything for accessibility */
+  @Prop() controlId: string = createNamespaceKey('cbp-button');
   /** The `name` attribute of the button, which is passed as part of formData (as a key) for the the pressed submit button. */
   @Prop() name: string;
   /** The `value` attribute of the button, which is passed as part of formData (as a value) for the the pressed submit button. */
@@ -52,18 +54,23 @@ export class CbpButton {
 
   /** Specifies if the button is pressed and results in `aria-pressed="true"` being placed on the button when true. Only valid on actual `button` elements. */
   @Prop() pressed: boolean;
+  
   /** 
    * Specifies if a controlled UI widget is expanded and results in `aria-pressed="true"` being placed on the button when true.
    * This property is usually used for progressive disclosure patterns such as accordions, menus, expand/collapse, etc., where
    * focus remains on the control after the user action.
    */
   @Prop() expanded: boolean;
-  /** Specifies the DOM element that the button controls and results in the `aria-controls` attribute
+
+  /** 
+   * Specifies the DOM element that the button controls and results in the `aria-controls` attribute
    * rendered on the button with the specified value.
    */
   @Prop() controls: string;
+  
   /* ??? */
   //@Prop() controlProp: "pressed" | "expanded";
+  
   /** The property on the target element being toggled by the button/control. */
   @Prop() targetProp: string; // A prop on the controlled element such as "open"
 
@@ -71,9 +78,6 @@ export class CbpButton {
    * or a sufficiently unique label. This text overrides the default label and is not additive to it.
    */
   @Prop() accessibilityText: string;
-
-  /* @Internal Specifies that a button should not be keyboard navigable by setting its tabindex to -1. This property should only be used in very specific cases. */
-  //@Prop() pointerOnly: boolean;
 
   /** Marks the rendered button/link in a disabled state when specified. */
   @Prop({reflect: true}) disabled: boolean;
@@ -102,6 +106,7 @@ export class CbpButton {
       // Toggle the prop it controls
       if (this.controlTarget) {
         this.controlTarget[this.targetProp] = !this.controlTarget[this.targetProp];
+        this.host.expanded = this.controlTarget[this.targetProp]
       } 
       else {
         console.warn('cbp-button configuration error: the control target referenced by ID by the `control` property could not be found.');
@@ -154,6 +159,13 @@ export class CbpButton {
       }
     }
 
+    if (this.button) {
+      this.button.getAttribute('id')
+        ? this.controlId = this.button.getAttribute('id')
+        : this.button.setAttribute('id', `${this.controlId}`);
+      if (this.disabled) this.button.setAttribute('disabled', '');
+    }
+
     setCSSProps(this.button, {
       'min-width': this.width,
       'min-height': this.height,
@@ -190,6 +202,7 @@ export class CbpButton {
             target,
           };
 
+    // slotted control
     if (this.host.querySelector('[slot=cbp-button-custom]')) {
       if (this.disabled) this.button.setAttribute("disabled",'');
       return (
@@ -198,6 +211,8 @@ export class CbpButton {
         </Host>
       );
     } 
+
+    // rendered button
     else if (this.tag === 'button') {
       return (
         <Host onClick={(e) => this.handleClick(e)}>
@@ -216,6 +231,8 @@ export class CbpButton {
         </Host>
       );
     } 
+
+    // rendered anchor/link with button styles
     else {
       return (
         <Host onClick={(e) => this.handleClick(e)}>