+++ packages/web-components/src/components/cbp-container/cbp-container.tsx
@@ -1,20 +1,30 @@
 import { Component, Prop, Element, Host, h } from '@stencil/core';
 import { setCSSProps } from '../../utils/utils';
 
+/**
+ * @slot - Any content or layout components may be slotted within the container.
+ */
 @Component({
   tag: 'cbp-container',
   styleUrl: 'cbp-container.scss',
 })
 export class CbpContainer {
+  
   @Element() host: HTMLElement;
 
+  /** Specifies the CSS background of the parent container, which could be a solid color, an image, a gradient, or any combination (or multiples) achievable via the CSS `background` property. */
   @Prop() background: string;
+  
+  /** Specifies the text color for content within the container, since it could be on any background. */
   @Prop() textColor: string;
+  
+  /** Specifies the width of the inner container area. */
   @Prop() width: string;
-  @Prop() padding: string;
+  
+  //@Prop() padding: string;
+  
   /** Supports adding inline styles as an object */
   @Prop() sx: any = {};
-
 
   componentWillLoad() {
     if (typeof this.sx == 'string') {
@@ -28,7 +38,7 @@ export class CbpContainer {
       ...this.sx,
     });
   }
-  
+ 
 
   render() {
     return (