+++ packages/web-components/src/stories/templates.stories.tsx
@@ -42,35 +42,26 @@ const InternalTemplate = ({ isLoggedIn, username }) => {
             <li>
               <cbp-button color="secondary" fill="ghost" context="dark-always">
                 <cbp-icon name="book"></cbp-icon>
-                <cbp-hide
-                  hide-at="max-width: 64em"
-                  sx='{"margin-left":"var(--cbp-space-2x)"}'
-                >App Directory</cbp-hide>
+                <cbp-hide visually-hide-at="max-width: 64em">App Directory</cbp-hide>
               </cbp-button>
             </li>
             <li>
               <cbp-button color="secondary" fill="ghost" context="dark-always">
                 <cbp-icon name="comment"></cbp-icon>  
-                <cbp-hide
-                  visually-hide-at="max-width: 64em"
-                  sx='{"margin-left":"var(--cbp-space-2x)"}'
-                >Feedback</cbp-hide>
+                <cbp-hide visually-hide-at="max-width: 64em">Feedback</cbp-hide>
               </cbp-button>
             </li>
             <li>
               <cbp-button color="secondary" fill="ghost" context="dark-always">
                 <cbp-icon name="user"></cbp-icon>
-                <cbp-hide
-                  visually-hide-at="max-width: 64em"
-                  sx='{"margin-left":"var(--cbp-space-2x)"}'
-                >${username}</cbp-hide>
+                <cbp-hide visually-hide-at="max-width: 64em">${username}</cbp-hide>
               </cbp-button>
             </li>
             `
             : `
             <li>
               <cbp-button tag="a" href="#" color="secondary" fill="ghost" context="dark-always">
-              <cbp-icon name="right-to-bracket" sx='{"margin-right":"var(--cbp-space-2x)"}'></cbp-icon>
+              <cbp-icon name="right-to-bracket"></cbp-icon>
               Login
               </cbp-button>
             </li>
@@ -175,35 +166,26 @@ const Internal2ColumnTemplate = ({ isLoggedIn, username, contentGridSize, sideba
             <li>
               <cbp-button color="secondary" fill="ghost" context="dark-always">
                 <cbp-icon name="book"></cbp-icon>
-                <cbp-hide
-                  hide-at="max-width: 64em"
-                  sx='{"margin-left":"var(--cbp-space-2x)"}'
-                >App Directory</cbp-hide>
+                <cbp-hide visually-hide-at="max-width:64em">App Directory</cbp-hide>
               </cbp-button>
             </li>
             <li>
               <cbp-button color="secondary" fill="ghost" context="dark-always">
                 <cbp-icon name="comment"></cbp-icon>  
-                <cbp-hide
-                  visually-hide-at="max-width: 64em"
-                  sx='{"margin-left":"var(--cbp-space-2x)"}'
-                >Feedback</cbp-hide>
+                <cbp-hide visually-hide-at="max-width:64em">Feedback</cbp-hide>
               </cbp-button>
             </li>
             <li>
               <cbp-button color="secondary" fill="ghost" context="dark-always">
                 <cbp-icon name="user"></cbp-icon>
-                <cbp-hide
-                  visually-hide-at="max-width: 64em"
-                  sx='{"margin-left":"var(--cbp-space-2x)"}'
-                >${username}</cbp-hide>
+                <cbp-hide visually-hide-at="max-width:64em">${username}</cbp-hide>
               </cbp-button>
             </li>
             `
             : `
             <li>
               <cbp-button tag="a" href="#" color="secondary" fill="ghost" context="dark-always">
-              <cbp-icon name="right-to-bracket" sx='{"margin-right":"var(--cbp-space-2x)"}'></cbp-icon>
+              <cbp-icon name="right-to-bracket"></cbp-icon>
               Login
               </cbp-button>
             </li>
@@ -381,35 +363,26 @@ const InternalCardsLayoutTemplate = ({ isLoggedIn, username, numberOfCards, card
             <li>
               <cbp-button color="secondary" fill="ghost" context="dark-always">
                 <cbp-icon name="book"></cbp-icon>
-                <cbp-hide
-                  hide-at="max-width: 64em"
-                  sx='{"margin-left":"var(--cbp-space-2x)"}'
-                >App Directory</cbp-hide>
+                <cbp-hide visually-hide-at="max-width:64em">App Directory</cbp-hide>
               </cbp-button>
             </li>
             <li>
               <cbp-button color="secondary" fill="ghost" context="dark-always">
                 <cbp-icon name="comment"></cbp-icon>  
-                <cbp-hide
-                  visually-hide-at="max-width: 64em"
-                  sx='{"margin-left":"var(--cbp-space-2x)"}'
-                >Feedback</cbp-hide>
+                <cbp-hide visually-hide-at="max-width:64em">Feedback</cbp-hide>
               </cbp-button>
             </li>
             <li>
               <cbp-button color="secondary" fill="ghost" context="dark-always">
                 <cbp-icon name="user"></cbp-icon>
-                <cbp-hide
-                  visually-hide-at="max-width: 64em"
-                  sx='{"margin-left":"var(--cbp-space-2x)"}'
-                >${username}</cbp-hide>
+                <cbp-hide visually-hide-at="max-width:64em">${username}</cbp-hide>
               </cbp-button>
             </li>
             `
             : `
             <li>
               <cbp-button tag="a" href="#" color="secondary" fill="ghost" context="dark-always">
-              <cbp-icon name="right-to-bracket" sx='{"margin-right":"var(--cbp-space-2x)"}'></cbp-icon>
+              <cbp-icon name="right-to-bracket"></cbp-icon>
               Login
               </cbp-button>
             </li>